+++ bridge-api/src/main/java/ccc/bridgeapi/BridgeApiApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgeapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgeApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgeApiApplication.class, args);
+    }
+
+}
