+++ bridge-api/src/main/java/ccc/bridgeapi/api/TestController.java
@@ -1,0 +1,15 @@
+package ccc.bridgeapi.api;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/test")
+public class TestController {
+
+    @GetMapping
+    public String helloWorld() {
+        return "Hello, World!";
+    }
+}
