+++ bridge-api/src/main/resources/application.yml
