+++ bridge-api/src/test/java/ccc/bridgeapi/BridgeApiApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgeapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgeApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
