+++ bridge-batch/src/main/java/ccc/bridgebatch/BridgeBatchApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgebatch;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgeBatchApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgeBatchApplication.class, args);
+    }
+
+}
