+++ bridge-batch/src/test/java/ccc/bridgebatch/BridgeBatchApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgebatch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgeBatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
