+++ bridge-domain/src/main/java/ccc/bridgedomain/BridgeDomainApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgedomain;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgeDomainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgeDomainApplication.class, args);
+    }
+
+}
