+++ bridge-domain/src/test/java/ccc/bridgedomain/BridgeDomainApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgedomain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgeDomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
