+++ bridge-infra/src/main/java/ccc/bridgeinfra/BridgeInfraApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgeinfra;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgeInfraApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgeInfraApplication.class, args);
+    }
+
+}
