+++ bridge-infra/src/test/java/ccc/bridgeinfra/BridgeInfraApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgeinfra;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgeInfraApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
