+++ bridge-push/src/main/java/ccc/bridgepush/BridgePushApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgepush;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgePushApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgePushApplication.class, args);
+    }
+
+}
