+++ bridge-push/src/test/java/ccc/bridgepush/BridgePushApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgepush;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgePushApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
