+++ bridge-statistics/src/main/java/ccc/bridgestatistics/BridgeStatisticsApplication.java
@@ -1,0 +1,13 @@
+package ccc.bridgestatistics;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BridgeStatisticsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BridgeStatisticsApplication.class, args);
+    }
+
+}
