+++ bridge-statistics/src/test/java/ccc/bridgestatistics/BridgeStatisticsApplicationTests.java
@@ -1,0 +1,13 @@
+package ccc.bridgestatistics;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BridgeStatisticsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
