+++ build.gradle
@@ -4,27 +4,107 @@ plugins {
 	id 'java'
 }
 
-group = 'ccc'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+subprojects {
+	apply plugin: 'java'
+	apply plugin: 'org.springframework.boot'
+	apply plugin: 'io.spring.dependency-management'
 
-configurations {
-	compileOnly {
-		extendsFrom annotationProcessor
+	group = 'ccc'
+	version = '0.0.1-SNAPSHOT'
+	sourceCompatibility = '11'
+
+	configurations {
+		compileOnly {
+			extendsFrom annotationProcessor
+		}
+	}
+
+	repositories {
+		mavenCentral()
+	}
+
+	dependencies {
+		compileOnly 'org.projectlombok:lombok'
+		annotationProcessor 'org.projectlombok:lombok'
+
+		testCompileOnly 'org.projectlombok:lombok'
+		testAnnotationProcessor 'org.projectlombok:lombok'
+
+		testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	}
+
+	ext {
+		springCloudVersion = "2021.0.1"
+	}
+
+	dependencyManagement {
+		imports {
+			mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+		}
+	}
+
+	tasks.named('test') {
+		useJUnitPlatform()
+	}
+
+}
+
+project(':bridge-batch') {
+	dependencies {
+		implementation 'org.springframework.boot:spring-boot-starter-batch'
+		implementation 'org.springframework.boot:spring-boot-starter-quartz'
 	}
 }
 
-repositories {
-	mavenCentral()
+project(':bridge-push') {
+	dependencies {
+		implementation 'org.springframework.boot:spring-boot-starter-amqp'
+	}
 }
 
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+project(':bridge-statistics') {
+	dependencies {
+		implementation project(path: ':bridge-domain', configuration: 'default')
+		implementation 'org.springframework.boot:spring-boot-starter-web'
+	}
 }
 
-tasks.named('test') {
-	useJUnitPlatform()
+project(':bridge-infra') {
+	bootJar {
+		enabled = false
+	}
+
+	jar {
+		enabled = true
+	}
+
+	dependencies {
+		implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
+		// todo :: amqp import, aws config
+	}
+}
+
+
+project(':bridge-domain') {
+	bootJar {
+		enabled = false
+	}
+
+	jar {
+		enabled = true
+	}
+
+	dependencies {
+		implementation project(path: ':bridge-infra', configuration: 'default')
+		// todo ::jpa, redis import
+	}
+}
+
+
+project(':bridge-api') {
+	dependencies {
+		implementation project(path: ':bridge-domain', configuration: 'default')
+		implementation 'org.springframework.boot:spring-boot-starter-web'
+	}
 }
