--- src/main/java/ccc/bridge/BridgeApplication.java
@@ -1,11 +1,0 @@
-package ccc.bridge;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BridgeApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(BridgeApplication.class, args);
-    }
-}
