--- src/main/java/ccc/bridge/controller/TestController.java
@@ -1,15 +1,0 @@
-package ccc.bridge.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/test")
-public class TestController {
-
-    @GetMapping
-    public String helloWorld() {
-        return "Hello, World!";
-    }
-}
