--- src/test/java/ccc/bridge/BridgeApplicationTests.java
@@ -1,13 +1,0 @@
-package ccc.bridge;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BridgeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
